requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==24.3.7
gast==0.4.0
google-auth==2.28.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.6
MarkupSafe==2.1.5
mega.py==1.0.8
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pathlib==1.0.1
pillow==10.2.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.20.0
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
tenacity==5.1.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing-extensions==4.10.0
urllib3==2.2.1
werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
